+++ src/plugin/gui/gui_prefs.cpp
@@ -55,7 +55,7 @@ prefs::Descriptor SettingDescriptor(GuiSetting setting) {
                     },
                 .default_value = (s64)k_default_gui_width,
                 .gui_label = "Window width",
-                .long_description = "The width of the main window.",
+                .long_description = "The width and scaling of Floe's window.",
             };
         case GuiSetting::Count: PanicIfReached();
     }